Cleanup renderer-side action dispatching
