+++ packages/actions/package.json
@@ -18,5 +18,8 @@
     "@nteract/types": "^5.1.1-alpha.0",
     "immutable": "^4.0.0-rc.12",
     "rx-jupyter": "^5.5.1-alpha.0"
+  },
+  "devDependencies": {
+    "utility-types": "^3.10.0"
   }
 }
